profiles/requirements.txt
django
djangorestframework
firebase-admin
python-decouple
